app/requirements.txt
Django==1.11.3
django-filter==1.0.4
django-versatileimagefield==1.7.1
djangorestframework==3.6.3
olefile==0.44
Pillow==4.0.0
Pygments==2.2.0
PyMySQL==0.7.11
pytz==2017.2